refactor(ai-connect): extract AiConnectPanel from wizard (no modal chrome)
Replaces the standalone AiConnectWizard modal with a content-only panel
that will live inside ConnectedAiClientsModal's Connect tab.
